requirements.txt
ag2[openai]
Pillow
